docs: added javadoc comments

+++ src/main/java/com/kauz/TicketRapport/controllers/TicketController.java
@@ -7,23 +7,40 @@ import com.kauz.TicketRapport.models.Ticket;
 import com.kauz.TicketRapport.models.filters.TicketFilter;
 import com.kauz.TicketRapport.models.templates.ChecklistItemTemplate;
 import com.kauz.TicketRapport.models.templates.ChecklistTemplate;
-import jakarta.annotation.security.RolesAllowed;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.*;
 
-import java.lang.reflect.Array;
-import java.util.*;
+import java.util.HashSet;
+import java.util.Objects;
+import java.util.Set;
+import java.util.UUID;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 /**
  * A controller for handling all requests pertaining to ticket data.
  */
 @Controller
 public class TicketController extends BaseController {
+
+    /**
+     * Get handler for the ticket list page.
+     * Fetches a list of filtered and sorted ticket entries to display.
+     * If the user is an admin, displays all tickets,
+     * otherwise display only the tickets that are assigned to the current user.
+     *
+     * @param model the model containing the relevant view data.
+     * @param id the id of the ticket to fetch, if present will display the ticket details, otherwise displays a list of tickets.
+     * @param referer a string representing the page the user came from / to return to.
+     * @param search the string to filter the tickets by.
+     * @param clientId the client id to filter the tickets by.
+     * @param statusId the status id to filter the tickets by.
+     * @param sort a comma separated string containing the sort order.
+     * @param page the current page as an integer.
+     * @param asc whether the sort order is ascending.
+     * @return a reference to a tickets Thymeleaf template.
+     */
     @GetMapping("/tickets")
     public String getIndex(Model model,
                            @RequestParam(required = false) UUID id,
@@ -63,8 +80,21 @@ public class TicketController extends BaseController {
         return "tickets/details";
     }
 
+    /**
+     * A post handler to update the status of a ticket. Should only be available to admins.
+     * Opens or closes a ticket depending on the args.
+     *
+     * @param id the id of the ticket to update.
+     * @param action the type of update to do (close or open).
+     * @param referer a string representing the page the user came from / to return to.
+     * @return a reference to a Thymeleaf template.
+     */
     @RequestMapping(value = "/tickets/status", method = RequestMethod.POST)
     public String status(@RequestParam UUID id, @RequestParam String action, @RequestParam(required = false) String referer) {
+        String returnString = referer.equals("home") ? "redirect:/" : "redirect:/tickets";
+
+        if (authUser.getUser().isAdmin()) return returnString;
+
         Ticket ticket = unitOfWork.getTicketService().find(Ticket.class, id);
         if (Objects.equals(action, "close")) {
             ticket.setStatus(unitOfWork.getStatusService().find("Closed"));
@@ -72,12 +102,24 @@ public class TicketController extends BaseController {
             ticket.setStatus(unitOfWork.getStatusService().find("In Progress"));
         }
         unitOfWork.getTicketService().update(ticket);
-        if (Objects.equals(referer, "home")) {
-            return "redirect:/";
-        }
-        return "redirect:/tickets";
+
+        return returnString;
     }
 
+    /**
+     * A get handler for a learner to update a ticket from the details page.
+     * Checks that the data is valid and saves the updated ticket in the database.
+     * Updates the protocol, solution, work time and checklist item status.
+     * Optionally also sets the ticket status to complete.
+     *
+     * @param id the id of the ticket to update.
+     * @param referer a string representing the page the user came from / to return to.
+     * @param entry the ticket entry to update.
+     * @param result information about the data binding.
+     * @param action the type of update to do (whether to mark the ticket as completed).
+     * @param checklistItems a comma separated string of ids for the checklist items that are marked as completed.
+     * @return a reference to a Thymeleaf template.
+     */
     @RequestMapping(value = "/tickets/update", method = RequestMethod.POST)
     public String update(@RequestParam UUID id,
                          @RequestParam(required = false) String referer,
@@ -115,6 +157,12 @@ public class TicketController extends BaseController {
         return "redirect:/tickets?id=" + id + "&referer=" + referer;
     }
 
+    /**
+     * A get handler for the form to create a new ticket.
+     *
+     * @param model the model containing the relevant view data.
+     * @return a reference to the Thymeleaf template of the ticket creation form.
+     */
     @GetMapping("/tickets/create")
     public String create(Model model) {
         super.addBaseAttributes(model);
@@ -126,6 +174,21 @@ public class TicketController extends BaseController {
         return "tickets/create";
     }
 
+    /**
+     * A post handler for creating a new ticket entry.
+     * Checks that the data is valid and saves a new ticket in the database.
+     * Also creates new checklist items and templates if necessary.
+     *
+     * @param checkboxes a comma separated string of values for the checklist items (checked or not).
+     * @param descriptions a comma separated string of descriptions for the checklist items.
+     * @param ids a comma separated list of ids for the checklist items.
+     * @param saveTemplate a boolean check for if a template of the checklist should be saved.
+     * @param templateName a string representing the name of the checklist template should it be saved.
+     * @param entry the ticket entry to persist.
+     * @param result information about the data binding.
+     * @param model the model containing the relevant view data.
+     * @return a reference to a Thymeleaf template.
+     */
     @RequestMapping(value = "/tickets/create", method = RequestMethod.POST)
     public String create(@RequestParam String checkboxes,
                          @RequestParam String descriptions,
@@ -152,12 +215,28 @@ public class TicketController extends BaseController {
         return "redirect:/tickets";
     }
 
+    /**
+     * A get handler to fetch checklist item templates from the database.
+     * Adds the template to the model and returns a fragment reference for displaying the data.
+     * For adding a checklist to a ticket from a template.
+     *
+     * @param id the id of the template entry.
+     * @param model the model containing the relevant view data (a stream of checklist item templates).
+     * @return a reference to the checklist Thymeleaf fragment.
+     */
     @GetMapping("/tickets/getTemplate")
     public String checklistTemplate(@RequestParam("id") UUID id, Model model) {
         model.addAttribute("items", unitOfWork.getChecklistItemTemplateService().findByTemplate(id));
         return "fragments/checklist :: checklist";
     }
 
+    /**
+     * A get handler for the form to edit a ticket.
+     *
+     * @param model the model containing the relevant view data.
+     * @param id the id of the ticket to edit.
+     * @return a reference to the ticket edit Thymeleaf template.
+     */
     @GetMapping("/tickets/edit")
     public String edit(Model model, @RequestParam UUID id) {
         super.addBaseAttributes(model);
@@ -169,12 +248,27 @@ public class TicketController extends BaseController {
         return "tickets/edit";
     }
 
+    /**
+     * A post handler for editing a ticket. Should only be available to admins.
+     * Checks that the data is valid and updates the ticket in the database.
+     * Also updates the checklist items and if necessary creates new checklist item entries.
+     *
+     * @param id the id of the ticket to update.
+     * @param checkboxes a comma separated list of the values of the checklist items (checked or not).
+     * @param descriptions a comma separated list of descriptions of the checklist items.
+     * @param ids a comma separated list of the ids of the checklist items.
+     * @param entry the new ticket data.
+     * @param result information about the data binding.
+     * @param model the model containing the relevant view data.
+     * @return a reference to a Thymeleaf template.
+     */
     @RequestMapping(value = "/tickets/edit", method = RequestMethod.POST)
     public String edit(@RequestParam UUID id,
                        @RequestParam String checkboxes,
                        @RequestParam String descriptions,
                        @RequestParam String ids,
                        @ModelAttribute Ticket entry, BindingResult result, Model model) {
+        if (!authUser.getUser().isAdmin()) return "redirect:/tickets";
         if (result.hasErrors()) {
             super.addBaseAttributes(model);
             model.addAttribute("entry", entry);
@@ -194,6 +288,16 @@ public class TicketController extends BaseController {
         updateChecklist(entry, checkboxes, descriptions, ids, false, null);
     }
 
+    /**
+     * Updates the ticket checklist items and optionally saves the new checklist as a template.
+     *
+     * @param entry the ticket to update.
+     * @param checkboxes a comma separated list of the values of the checklist items (checked or not).
+     * @param descriptions a comma separated list of descriptions of the checklist items.
+     * @param ids a comma separated list of the ids of the checklist items.
+     * @param saveTemplate a boolean check for if a template of the checklist should be saved.
+     * @param templateName a string representing the name of the checklist template should it be saved.
+     */
     private void updateChecklist(Ticket entry, String checkboxes, String descriptions, String ids, boolean saveTemplate, String templateName) {
         String[] checkArray = checkboxes.split(";;");
         String[] descArray = descriptions.split(";;");
@@ -247,6 +351,13 @@ public class TicketController extends BaseController {
         }
     }
 
+    /**
+     * A get handler for the form to delete a ticket entry.
+     *
+     * @param model the model containing the relevant view data.
+     * @param id the id of the ticket to delete.
+     * @return a reference to the ticket delete Thymeleaf template.
+     */
     @GetMapping("/tickets/delete")
     public String delete(Model model, @RequestParam UUID id) {
         super.addBaseAttributes(model);
@@ -254,9 +365,17 @@ public class TicketController extends BaseController {
         return "tickets/delete";
     }
 
+    /**
+     * A post handler for deleting a ticket entry. Should only be available to admins
+     * Checks that the data is valid and deletes the ticket from the database.
+     * @param id the id of the ticket to delete.
+     * @param entry the ticket entry to remove.
+     * @param result information about the data binding.
+     * @return a reference to the ticket list Thymeleaf template.
+     */
     @RequestMapping(value = "/tickets/delete", method = RequestMethod.POST)
     public String delete(@RequestParam UUID id, @ModelAttribute Ticket entry, BindingResult result) {
-        if (!result.hasErrors()) {
+        if (!result.hasErrors() && authUser.getUser().isAdmin()) {
             unitOfWork.getTicketService().delete(Ticket.class, entry);
         }
         return "redirect:/tickets";
